Create Binomial_Poisson
A random variable X with a Binomial Distribution and parameters (n, p) when n tends to infinity, X approaches a Poisson distribution. 
We will now computationally demonstrate how this statement becomes true: 
Consider the following problem situation for example: Let's consider an example where we flip a coin n times and count the number of heads. 
Initially, we will define a reasonably large value for n and perform this experiment using a binomial distribution. Then, we will reduce n considerably and show that, as n tends to infinity, the binomial distribution increasingly resembles the Poisson distribution.